rename eventStore to cache
